🚚 [Notebook] Separate `DeportedOutputsView` in a dedicated file
